reorganized resolver specs
